% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_unit_functions.R
\name{chilling_degree_unit}
\alias{chilling_degree_unit}
\title{Calculate Daily Chilling Degree Units (CDU)}
\usage{
chilling_degree_unit(temp, chilling_threshold = 7)
}
\arguments{
\item{temp}{A numeric vector of daily mean temperatures (in degrees Celsius).}

\item{chilling_threshold}{A numeric value for the threshold (default is 7°C).}
}
\value{
A numeric vector of daily CDU values.
}
\description{
Computes the daily Chilling Degree Unit (CDU) to assess cold exposure. It calculates the difference
between a specified chilling threshold and the daily temperature, only when the temperature is below that threshold.
If the temperature is above the threshold, the chilling unit is zero.
}
\examples{
# For a vector of daily temperatures
temps <- c(4, 8, 6)
chilling_degree_unit(temps)

}
