% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_stage.R
\name{corn_growth_stage}
\alias{corn_growth_stage}
\title{Determine Corn Growth Stage Based on Cumulative GDUs}
\usage{
corn_growth_stage(
  cum_gdu,
  thresholds = c(VE = 100, V6 = 300, VT = 900, R1 = 1000, R6 = 2000)
)
}
\arguments{
\item{cum_gdu}{A numeric vector of cumulative growing degree units.}

\item{thresholds}{A named numeric vector of thresholds. The defaults are
\code{c(VE = 100, V6 = 300, VT = 900, R1 = 1000, R6 = 2000)}.}
}
\value{
A character vector indicating the corn growth stage.
}
\description{
This function assigns the corn growth stage from a numeric vector of cumulative growing degree
units (GDU), using default thresholds based on widely used agronomic practices (e.g., McMaster & Wilhelm, 1997)
and the TidyAgronomy framework.
}
\details{
The default stages are:
\describe{
\item{Pre-VE}{cum_gdu < 100.}
\item{VE}{100 ≤ cum_gdu < 300.}
\item{V6}{300 ≤ cum_gdu < 900.}
\item{VT}{900 ≤ cum_gdu < 1000.}
\item{R1}{1000 ≤ cum_gdu < 2000.}
\item{R6}{cum_gdu ≥ 2000.}
}
}
\references{
McMaster, G. S., & Wilhelm, W. W. (1997). Growing degree-days: one equation, two interpretations.
\emph{Agricultural and Forest Meteorology, 87}(4), 291–300.
}
