% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_data_readers.R
\name{easy_era5}
\alias{easy_era5}
\title{easy readers for TidyAgronomy
Do you want to contribute? Submit a pull request!}
\usage{
easy_era5(filePath)
}
\description{
easy readers for TidyAgronomy
Do you want to contribute? Submit a pull request!
}
