% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_unit_functions.R
\name{growing_degree_unit}
\alias{growing_degree_unit}
\title{Growing Degree & Stress Calculations
References: https://products.climate.ncsu.edu/ag/corn/about/}
\usage{
growing_degree_unit(value, baseline, upper_bound = NULL)
}
\description{
Growing Degree & Stress Calculations
References: https://products.climate.ncsu.edu/ag/corn/about/
}
