% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_compressions.R
\name{mean_daily}
\alias{mean_daily}
\title{Compress hourly weather using a variety of methods.
The first is converting hourly to "mean daily"}
\usage{
mean_daily(.data, location_id_column, season_column, time_column, temp_column)
}
\description{
TODO: citation
the user is left to fill in the details if they wish to change the aggregation method
}
