% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_era5}
\alias{read_era5}
\title{easy readers for tidyagronomy
Do you want to contribute? Submit a pull request!}
\usage{
read_era5(filePath)
}
\description{
easy readers for tidyagronomy
Do you want to contribute? Submit a pull request!
}
